requirements.txt
Chameleon==2.13-1
Mako==0.9.0
MarkupSafe==0.18
PasteDeploy==1.5.0
Pygments==1.6
SPARQLWrapper==1.5.2
SQLAlchemy==0.9.3
WebOb==1.3.1
WebTest==2.0.9
alembic==0.6.2
argparse==1.2.1
beautifulsoup4==4.3.2
clld>=0.15.2
clldmpg
colander==0.9.9
distribute==0.6.24
ecdsa==0.10
html5lib==0.95
isodate==0.4.9
paginate==0.4.1
path.py==5.0
purl==0.8
pyparsing==1.5.7
pyramid==1.5b1
pyramid-debugtoolbar==1.0.9
pyramid-exclog==0.7
pyramid-mako==0.3.1
pyramid-tm==0.7
python-dateutil==2.2
pytz==2013.8
rdflib==4.0.1
repoze.lru==0.6
requests==2.0.1
six==1.4.1
transaction==1.4.1
translationstring==1.1
unicsv==1.0.0
venusian==1.0a8
waitress==0.8.7
webassets==0.9
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.5
yuicompressor==2.4.8
zope.deprecation==4.0.2
zope.interface==4.0.5
zope.sqlalchemy==0.7.3
